Time: 2 ms (53.47%), Space: 44.9 MB (38.3%) - LeetHub

+++ 0152-maximum-product-subarray/0152-maximum-product-subarray.java
@@ -1,14 +1,15 @@
 class Solution {
     public int maxProduct(int[] nums) {
+
         if (nums.length == 0) return 0;
 
-        double max_so_far = nums[0];
-        double min_so_far = nums[0];
-        double result = max_so_far;
+        int max_so_far = nums[0];
+        int min_so_far = nums[0];
+        int result = max_so_far;
 
         for (int i = 1; i < nums.length; i++) {
             int curr = nums[i];
-            double temp_max = Math.max(
+            int temp_max = Math.max(
                 curr,
                 Math.max(max_so_far * curr, min_so_far * curr)
             );
